committing work done on order listing on single

+++ lib/core/connect_end/model/place_order_wallet_response_model/delivery_breakdown.dart
@@ -10,7 +10,7 @@ class DeliveryBreakdown {
   List<String>? productIds;
   List<String>? distributorIds;
   List<String>? distributorNames;
-  double? orderSubtotal;
+  int? orderSubtotal;
   String? locationId;
   String? stateCode;
   String? stateName;
